requirements.txt
flask==3.0.3
flask-cors==4.0.1
requests==2.32.3
openai==1.42.0
python-dotenv==1.0.1
pandas==2.2.2
numpy==2.0.1
matplotlib==3.9.1
schedule==1.2.2
urllib3==2.2.2